fix(v2): review 发现的序列布局错位 + bf16 CE 精度回退 + 保存/默认值
P0 —— tts_eos 与 codec_bos 被并进同一格，尾部整体错位一格
  推理侧（modeling:2224-2245 non-streaming 分支）分两段拼：
    ① text_emb(正文)+tts_eos 叠 codec_pad ×(n_body+1)
    ② tts_pad_embed          叠 emb(codec_bos)   ← codec_bos 独占一格
  官方 dataset.py:174/175 vs :190/191 同样是分两格。合成一格后，推理时模型会
  在训练中从未出现过的 (tts_pad, codec_bos) 组合上被要求吐第一帧 codec，而
  (tts_eos, codec_pad) 也从未共现——不报错，只表现为训完推理就崩。
  修完逐格核对：pos21 (tts_eos, codec_pad) / pos22 (tts_pad, codec_bos) 且
  label 起点正落在 pos22，codec_bos 预测第一帧、末帧预测 codec_eos。
  连带把 collate 缓冲从 max(lens)+8 改为 max(lens)+nb+4（原值在 nb=5 时零余量）。

P1 —— _compute_token_ce_loss 丢了 logits.float()
  被替换的 HF ForCausalLMLoss 第一行就是 logits.float()。训练脚本直接调 talker
  子模块，accelerate 的 autocast 只包顶层 forward，这里是裸 bf16，log_softmax
  在 bf16 下算会掉信噪比。

其余：
- --max_steps 提前结束时不落盘 → 保存抽成 _save()，break 前也调
- --attn 默认 flash_attention_2 → sdpa（NPU 上前者不可用，少踩一次）
- 训练用的 language 写进 config 的 v2_train_language：它决定 codec 前缀是
  think+lang(5 格) 还是 Auto 的 nothink(4 格)，而 generate_voice_design 默认
  Auto，不对齐就是静默失配
- 保存前删 model.safetensors.index.json：源目录若是分片 ckpt，index 会让新写的
  单文件失效，加载时静默读回原始权重
- talker(...) 传 use_cache=False；save_file 补 metadata={"format":"pt"}
- prepare_data.py 的 --device 补 NPU 提示（该步在 NPU 上必须显式传 npu:0）
- 清理 dataset_voicedesign.py 未使用的 import

+++ qwen_tts/core/models/modeling_qwen3_tts.py
@@ -104,8 +104,11 @@ def _compute_token_ce_loss(logits: torch.Tensor, labels: torch.Tensor, vocab_siz
     脚本若同时手工 shift，就变成双 shift——不报错，只表现为训不出东西。
     统一约定：**模型侧不 shift，由 collate 把 label 摆到因果前一格**。
     """
+    # .float() 不能省：HF 的 ForCausalLMLoss 第一行就是 logits.float()。
+    # 训练脚本直接调 talker 子模块，accelerate 的 autocast 只包顶层 forward，
+    # 这里拿到的是裸 bf16，log_softmax 在 bf16 下算会明显掉信噪比。
     return F.cross_entropy(
-        logits.reshape(-1, vocab_size),
+        logits.reshape(-1, vocab_size).float(),
         labels.reshape(-1),
         ignore_index=-100,
     )
